Triding_System/multi_bot_trading_system/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pandas==2.0.0
numpy==1.24.0
python-binance==1.0.17
alpaca-trade-api==3.2.0
requests==2.31.0
python-dotenv==1.0.0
ta==0.10.2
ccxt==2.77.0
